Move to submodule and fix a few nits
