+++ crates/ty_python_semantic/src/types/diagnostic.rs
@@ -7,13 +7,13 @@ use super::{
 };
 use crate::lint::{Level, LintRegistryBuilder, LintStatus};
 use crate::suppression::FileSuppressionId;
+use crate::types::LintDiagnosticGuard;
 use crate::types::function::KnownFunction;
 use crate::types::string_annotation::{
     BYTE_STRING_TYPE_ANNOTATION, ESCAPE_CHARACTER_IN_FORWARD_ANNOTATION, FSTRING_TYPE_ANNOTATION,
     IMPLICIT_CONCATENATED_STRING_TYPE_ANNOTATION, INVALID_SYNTAX_IN_FORWARD_ANNOTATION,
     RAW_STRING_TYPE_ANNOTATION,
 };
-use crate::types::{LintDiagnosticGuard, all_members};
 use crate::types::{SpecialFormType, Type, protocol_class::ProtocolClassLiteral};
 use crate::{Db, Module, ModuleName, Program, declare_lint};
 use itertools::Itertools;
@@ -22,6 +22,9 @@ use ruff_python_ast::{self as ast, AnyNodeRef};
 use ruff_text_size::{Ranged, TextRange};
 use rustc_hash::FxHashSet;
 use std::fmt::Formatter;
+pub(crate) use levenshtein::find_best_suggestion_for_unresolved_member;
+
+mod levenshtein;
 
 /// Registers all known type check lints.
 pub(crate) fn register_lints(registry: &mut LintRegistryBuilder) {
@@ -2172,186 +2175,4 @@ pub(super) fn hint_if_stdlib_submodule_exists_on_other_versions(
     ));
 
     add_inferred_python_version_hint_to_diagnostic(db, diagnostic, "resolving modules");
-}
-
-pub(super) fn find_best_suggestion_for_unresolved_member<'db>(
-    db: &'db dyn Db,
-    obj: Type<'db>,
-    unresolved_member: &str,
-) -> Option<ast::name::Name> {
-    find_best_suggestion(all_members(db, obj), unresolved_member)
-}
-
-/// The cost of a Levenshtein insertion, deletion, or substitution.
-///
-/// This is used instead of the conventional unit cost to give these differences a higher cost than
-/// casing differences, which CPython assigns a cost of 1.
-const MOVE_COST: usize = 2;
-
-fn find_best_suggestion(
-    options: impl IntoIterator<Item = ast::name::Name>,
-    unresolved_member: &str,
-) -> Option<ast::name::Name> {
-    if unresolved_member.is_empty() {
-        return None;
-    }
-
-    let mut best_suggestion = None;
-    for member in options {
-        let mut max_distance = (member.len() + unresolved_member.len() + 3) * MOVE_COST / 6;
-        if let Some((_, best_distance)) = best_suggestion {
-            if best_distance > 0 {
-                max_distance = max_distance.min(best_distance - 1);
-            }
-        }
-        let current_distance = levenshtein(unresolved_member, &member, max_distance);
-        let max_distance = (unresolved_member.len() + member.len() + 3) / 3;
-        if current_distance > max_distance {
-            continue;
-        }
-        if best_suggestion
-            .as_ref()
-            .is_none_or(|(_, best_score)| &current_distance < best_score)
-        {
-            best_suggestion = Some((member, current_distance));
-        }
-    }
-    best_suggestion.map(|(suggestion, _)| suggestion)
-}
-
-/// Determine the "cost" of converting `string_a` to `string_b`.
-fn substitution_cost(char_a: char, char_b: char) -> CharacterMatch {
-    if char_a == char_b {
-        return CharacterMatch::Exact;
-    }
-
-    if char_a
-        .to_lowercase()
-        .zip(char_b.to_lowercase())
-        .all(|(a, b)| a == b)
-    {
-        return CharacterMatch::CaseInsensitive;
-    }
-
-    CharacterMatch::None
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-enum CharacterMatch {
-    Exact,
-    CaseInsensitive,
-    None,
-}
-
-/// Returns the [Levenshtein edit distance] between strings `string_a` and `string_b`.
-/// Uses the [Wagner-Fischer algorithm] to speed up the calculation.
-///
-/// [Levenshtein edit distance]: https://en.wikipedia.org/wiki/Levenshtein_distance
-/// [Wagner-Fischer algorithm]: https://en.wikipedia.org/wiki/Wagner%E2%80%93Fischer_algorithm
-fn levenshtein(string_a: &str, string_b: &str, max_cost: usize) -> usize {
-    if string_a == string_b {
-        return 0;
-    }
-
-    let string_a_chars: Vec<_> = string_a.chars().collect();
-    let string_b_chars: Vec<_> = string_b.chars().collect();
-
-    let pre = string_a_chars
-        .iter()
-        .zip(string_b_chars.iter())
-        .take_while(|(a, b)| a == b)
-        .count();
-
-    let string_a_chars = &string_a_chars[pre..];
-    let string_b_chars = &string_b_chars[pre..];
-
-    let post = string_a_chars
-        .iter()
-        .rev()
-        .zip(string_b_chars.iter().rev())
-        .take_while(|(a, b)| a == b)
-        .count();
-
-    let mut string_a_chars = &string_a_chars[..string_a_chars.len() - post];
-    let mut string_b_chars = &string_b_chars[..string_b_chars.len() - post];
-
-    let mut string_a_len = string_a_chars.len();
-    let mut string_b_len = string_b_chars.len();
-
-    if string_a_len == 0 || string_b_len == 0 {
-        return MOVE_COST * (string_a_len + string_b_len);
-    }
-
-    // Prefer a shorter buffer
-    if string_b_chars.len() < string_a_chars.iter().len() {
-        std::mem::swap(&mut string_a_chars, &mut string_b_chars);
-        std::mem::swap(&mut string_a_len, &mut string_b_len);
-    }
-
-    if (string_b_len - string_a_len) * MOVE_COST > max_cost {
-        return max_cost + 1;
-    }
-
-    let mut row = vec![0; string_a_len];
-    for (i, v) in (MOVE_COST..MOVE_COST * (string_a_len + 1))
-        .step_by(MOVE_COST)
-        .enumerate()
-    {
-        row[i] = v;
-    }
-
-    let mut result = 0;
-    for bindex in 0..string_b_len {
-        let bchar = string_b_chars[bindex];
-        result = bindex * MOVE_COST;
-        let mut distance = result;
-        let mut minimum = std::usize::MAX;
-        for index in 0..string_a_len {
-            let substitute = distance + substitution_cost(bchar, string_a_chars[index]) as usize;
-            distance = row[index];
-            let insert_delete = result.min(distance) + MOVE_COST;
-            result = insert_delete.min(substitute);
-
-            row[index] = result;
-            if result < minimum {
-                minimum = result;
-            }
-        }
-
-        if minimum > max_cost {
-            return max_cost + 1;
-        }
-    }
-
-    result
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_bad_suggestions_do_not_trigger_for_small_names() {
-        let candidates = ["vvv", "mom", "w", "id", "pytho"].map(ast::name::Name::from); // # spellchecker:disable-line
-        for name in ["b", "v", "m", "py"] {
-            let suggestion = find_best_suggestion(candidates.clone(), name);
-            if let Some(suggestion) = suggestion {
-                panic!("Expected no suggestions for `{name}` but `{suggestion}` was suggested");
-            }
-        }
-    }
-
-    #[test]
-    fn test_levenshtein() {
-        let tests = [
-            // These are from the Levenshtein Wikipedia article, updated to match CPython's
-            // implementation (just doubling the score to accommodate the MOVE_COST)
-            ("kitten", "sitting", 6),
-            ("uninformed", "uniformed", 2),
-            ("flaw", "lawn", 4),
-        ];
-        for (a, b, want) in tests {
-            assert_eq!(levenshtein(a, b, std::usize::MAX), want);
-        }
-    }
 }
